+++ CHANGELOG.rst
@@ -17,6 +17,7 @@ Version 0.2
   - Day of week is now a feature in dataset.
   - Add win percentage and pythagorean expectation to dataset.
   - switch to joblib for sklearn model persistence
+  - Set DataStore default to Database
 
 Version 0.1
 -----------